# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"githubId": {
						"type": "string",
						"nullable": true
					},
					"accessToken": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"password": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"email",
					"username",
					"githubId",
					"accessToken",
					"name",
					"password",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUserRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_CreateUserRequest_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "backend-repo",
		"version": "1.0.0",
		"description": "Backend service with Swagger documentation and Prisma ORM",
		"contact": {}
	},
	"paths": {
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 1,
											"email": "user@example.com",
											"username": "username",
											"githubId": null,
											"accessToken": null,
											"name": null,
											"password": null,
											"createdAt": "2025-02-17T15:37:49.950Z",
											"updatedAt": "2025-02-17T15:37:49.951Z"
										}
									}
								}
							}
						}
					},
					"201": {
						"description": "Created"
					}
				},
				"description": "Create a new user",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserRequest"
							}
						}
					}
				}
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetUserProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 1,
											"email": "user@example.com",
											"username": "username",
											"githubId": null,
											"accessToken": null,
											"name": null,
											"password": null,
											"createdAt": "2025-02-17T15:37:49.954Z",
											"updatedAt": "2025-02-17T15:37:49.954Z"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Get user profile by ID",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUserProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 1,
											"email": "updated@example.com",
											"username": "updated_username",
											"githubId": null,
											"accessToken": null,
											"name": "Updated Name",
											"password": null,
											"createdAt": "2025-02-17T15:37:49.954Z",
											"updatedAt": "2025-02-17T15:37:49.954Z"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Update user profile",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_CreateUserRequest_"
							}
						}
					}
				}
			}
		},
		"/users/me": {
			"get": {
				"operationId": "GetCurrentUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated"
					}
				},
				"description": "Get current authenticated user",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/users/auth/github/callback": {
			"get": {
				"operationId": "GithubCallback",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "GitHub OAuth callback",
				"security": [],
				"parameters": []
			}
		},
		"/users/logout": {
			"get": {
				"operationId": "Logout",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Logout current user",
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}